refactor: ApiError의 패키지 이동
