+++ fourtoon-cookie/src/pages/DiaryWritePage/WriteDoneButtonLayout/WriteDoneButtonLayout.tsx
@@ -7,7 +7,7 @@ import { Diary, DiaryStatus } from "../../../types/diary";
 import { useSelectedCharacterStore } from "../../../hooks/store/selectedCharacter";
 import { RootStackParamList } from "../../../constants/routing";
 import { API_STATUS } from "../../../constants/api";
-import { ApiError } from "../../../error/ApiError";
+import { ApiError } from "../../../types/error/ApiError";
 import handleError from "../../../error/errorhandler";
 import { GlobalErrorInfoType } from "../../../types/error";
 import Button from "../../../components/common/Button/Button";
